ci: 补齐审查剩余修复 — CI 质量门禁、clippy 门禁、环境诊断
- 新增 .github/workflows/ci.yml：push/PR 时前端 lint/type-check/build/vitest
  （ubuntu-latest）与 Rust clippy(-D warnings)/test（windows-latest，代码依赖 Windows API）
- Cargo.toml 增加 [lints.clippy] 策略：允许 too_many_arguments（DB/IPC 参数与
  IPC 载荷 1:1 对应），其余 lint 保持 deny；新增 npm run clippy 本地入口
- 修复 clippy 告警：repeat_n、clamp、去 identity map、冗余 cast、&PathBuf → &Path
- 新增 npm run doctor（scripts/doctor.mjs）：检查 Node/npm/Rust/Tauri CLI/
  WebView2/SQLite（文件头校验 + 可选 integrity_check）/媒体目录，异常附修复建议
- CLAUDE.md / README 补充 doctor、clippy 与 CI 说明

+++ src-tauri/src/tray.rs
@@ -204,7 +204,7 @@ fn windows_power_watch_loop() {
         lparam: LPARAM,
     ) -> LRESULT {
         if msg == WM_POWERBROADCAST {
-            let event = wparam as usize;
+            let event = wparam;
             if matches!(
                 event,
                 PBT_APMRESUMESUSPEND | PBT_APMRESUMEAUTOMATIC | PBT_APMRESUMECRITICAL
@@ -293,7 +293,7 @@ pub(crate) fn hook_tray_menu_cursor(window: &tauri::WebviewWindow) {
         _id: usize,
         _data: usize,
     ) -> LRESULT {
-        if msg == WM_SETCURSOR && (lparam as u32 & 0xFFFF) == HTCLIENT as u32 {
+        if msg == WM_SETCURSOR && (lparam as u32 & 0xFFFF) == HTCLIENT {
             return 1; // WebView2 manages the cursor in its client area.
         }
         DefSubclassProc(hwnd, msg, wparam, lparam)
